Moved saveconfig, loadconfig, defaultconfig to new file config.lua
